break "actor should run" logic into a bunch of separate funcs
