+++ pkg/actor/director_test.go
